requirements.txt
google-genai
scapy
pandas
python-dotenv
rich
pillow